requirements.txt
aws-cdk-lib==2.110.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.110.0a0
boto3==1.34.0
